requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
cloudinary==1.29.0
decouple==0.0.7
dj-database-url==0.5.0
Django==4.0.5
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-heroku==0.3.1
djangorestframework==3.13.1
heroku==0.1.4
idna==3.3
Pillow==9.1.1
psycopg2==2.9.3
python-dateutil==1.5
python-decouple==3.6
pytz==2022.1
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
